backend/requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
colorama==0.4.6
coverage==7.2.2
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pytest==7.2.2
pytest-cov==4.0.0
python-dotenv==1.0.0
pytz==2023.2
six==1.16.0
SQLAlchemy==2.0.7
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
